Update margins tranasaction model
