+++ racun-service/src/main/java/rs/edu/raf/banka/racun/repository/MarginTransakcijaRepository.java
@@ -1,7 +1,7 @@
 package rs.edu.raf.banka.racun.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import rs.edu.raf.banka.racun.model.MarginTransakcija;
+import rs.edu.raf.banka.racun.model.margins.MarginTransakcija;
 
 public interface MarginTransakcijaRepository extends JpaRepository<MarginTransakcija, Long> {
